Convert DebuggerTraceListener to IPackageService
